fonction div, erreur core dumped

+++ push_swap.h
@@ -6,12 +6,13 @@
 # include <stdlib.h>
 # include <string.h>
 # include <stdio.h>
+# include <limits.h>
 
 int ft_count(const char *s, char c);
 int ft_len(const char *s, char c, int i);
 char    **ft_split(const char *s, char c);
 int  ft_strlen(const char *s);
 
-int ft_atoi(const char *s);
+long ft_atoi(const char *s);
 
 #endif
